+++ setenv_sample.js
@@ -3,3 +3,5 @@ process.env.ADT_USER = "developer"
 process.env.ADT_PASS = "secret"
 process.env.ADT_TRANS = "NPLK900060"
 process.env.ADT_SYSTEMID = "NPL"
+// uncomment the following line to allow disrupting tests i.e. create and release transports, create and delete objects
+// process.env.ADT_ENABLE_ALL = "YES"